+++ src/components/toaster.tsx
@@ -1,12 +1,10 @@
 import * as React from 'react';
-import { setup } from 'goober';
+import { CSSAttribute } from 'goober';
 import CSS from 'csstype';
 
 import { useToaster } from '../core/use-toaster';
 import { ToastBar } from './toast-bar';
 import { ToastPosition, DefaultToastOptions } from '../core/types';
-
-setup(React.createElement);
 
 interface ToasterProps {
   position?: ToastPosition;
@@ -15,6 +13,36 @@ interface ToasterProps {
 
   toastOptions?: DefaultToastOptions;
 }
+
+const getPositionStyle = (
+  position: ToastPosition,
+  offset: number
+): React.CSSProperties => {
+  const top = position.includes('top');
+  const verticalStyle: CSSAttribute = top ? { top: 0 } : { bottom: 0 };
+  const horizontalStyle: CSSAttribute = position.includes('left')
+    ? {
+        left: 0,
+      }
+    : position.includes('right')
+    ? {
+        right: 0,
+      }
+    : {
+        left: 0,
+        pointerEvents: 'none',
+        right: 0,
+        justifyContent: 'center',
+      };
+  return {
+    display: 'flex',
+    position: 'fixed',
+    transition: 'all 230ms cubic-bezier(.21,1.02,.73,1)',
+    transform: `translateY(${offset * (top ? 1 : -1)}px)`,
+    ...verticalStyle,
+    ...horizontalStyle,
+  };
+};
 
 export const Toaster: React.FC<ToasterProps> = ({
   reverseOrder,
@@ -35,16 +63,25 @@ export const Toaster: React.FC<ToasterProps> = ({
       onMouseLeave={handlers.endPause}
     >
       {toasts.map((t) => {
+        const offset = handlers.calculateOffset(t.id, {
+          reverseOrder,
+        });
+        const positionStyle = getPositionStyle(position, offset);
+
         return (
-          <ToastBar
+          <div
             key={t.id}
-            onHeight={(height) => handlers.updateHeight(t.id, height)}
-            toast={t}
-            offset={handlers.calculateOffset(t.id, {
-              reverseOrder,
-            })}
-            position={position}
-          />
+            style={{
+              zIndex: t.visible ? 9999 : undefined,
+              ...positionStyle,
+            }}
+          >
+            <ToastBar
+              onHeight={(height) => handlers.updateHeight(t.id, height)}
+              toast={t}
+              position={position}
+            />
+          </div>
         );
       })}
     </div>
